// retailercart/node_service/user-profile-service/src/shared/enum/column-type.enum.ts
export enum SQLColumnType {
  INT = 'int',
  INT2 = 'int2',
  INT4 = 'int4',
  INT8 = 'int8',
  INTEGER = 'integer',
  TINY_INT = 'tinyint',
  SMALLINT = 'smallint',
  VARCHAR = 'varchar',
  BOOLEAN = 'boolean',
  CHAR = 'char',
  DATE_TIME = 'datetime',
  DECIMAL = 'decimal',
  BLOB = 'blob',
  TEXT = 'text',
}